ex02 progress

+++ ex01/RPN.hpp
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   RPN.hpp                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rchahban <rchahban@student.1337.ma>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/05/28 19:51:13 by rchahban          #+#    #+#             */
+/*   Updated: 2024/07/11 01:24:44 by rchahban         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef RPN_HPP
 # define RPN_HPP
 
@@ -8,13 +20,14 @@
 class RPN
 {
     private:
-        std::stack<int> numbers;
+        std::stack<double> numbers;
     public:
         RPN();
         ~RPN();
         RPN(const RPN& original);
         RPN& operator=(const RPN& original);
         void calculate(std::string& expression);
+		// std::string trim(std::string &str);
 };
